requirements.txt
# Core orchestration & LLMs
langgraph>=0.2.0
langchain>=0.2.5
langchain-openai>=0.1.7
openai>=1.30.0

# Data validation & typing
pydantic>=2.6.0

# SQL parsing & DB access
sqlglot>=18.0.0
SQLAlchemy>=2.0.25
psycopg[binary]>=3.1.18   # PostgreSQL (or use psycopg2-binary if you prefer)
PyMySQL>=1.1.0            # MySQL/MariaDB
aiosqlite>=0.19.0         # Async SQLite helper (sqlite3 is built-in)

# UI
gradio>=4.28.0

# Utilities
python-dotenv>=1.0.1
tenacity>=8.2.3           # robust retries for tools
orjson>=3.10.0            # fast JSON (optional but nice)
rich>=13.7.1              # pretty logs/trace in terminal

# Development & Testing
pytest>=8.2.0
pytest-asyncio>=0.23.8
pytest-cov>=5.0.0
ruff>=0.4.10
black>=24.4.2
mypy>=1.10.0
types-PyYAML>=6.0.12.20240808